improve: global refactoring
