+++ Components/Settings/styles/DeconnectionButtonStyle.js
