+++ Components/Users/styles/FindPlacesCardStyle.js
