+++ __tests__/Leave/components/LeaveButton_test.js
@@ -3,7 +3,7 @@ import { View } from "react-native";
 import enzyme, { shallow } from "enzyme";
 import { expect } from "chai";
 import ReactSixteenAdapter from "enzyme-adapter-react-16";
-import LeaveButton from "../../../Components/Leave/components/LeaveButton";
+import LeaveButton from "../../../Components/Leave/LeaveButton";
 
 enzyme.configure({ adapter: new ReactSixteenAdapter() });
