+++ __tests__/Login/components/InputLogin_test.js
@@ -3,7 +3,7 @@ import { View } from "react-native";
 import enzyme, { shallow } from "enzyme";
 import { expect } from "chai";
 import ReactSixteenAdapter from "enzyme-adapter-react-16";
-import InputLogin from "../../../Components/Login/components/InputLogin";
+import InputLogin from "../../../Components/Login/InputLogin";
 
 enzyme.configure({ adapter: new ReactSixteenAdapter() });
