+++ __tests__/Places/PlacesScreen.test.js
@@ -5,8 +5,8 @@ import React from "react";
 import { expect } from "chai";
 import enzyme, { shallow } from "enzyme";
 import ReactSixteenAdapter from "enzyme-adapter-react-16";
-import PlacesScreen from "../../Components/Profile/Places/PlacesScreen";
-import FetchPlacesButton from "../../Components/Profile/Places/components/FetchPlacesButton";
+import PlacesScreen from "../../views/Places/PlacesScreen";
+import FetchPlacesButton from "../../Components/Places/FetchPlacesButton";
 
 enzyme.configure({ adapter: new ReactSixteenAdapter() });
