+++ __tests__/Users/components/FindPlacesCard.test.js
@@ -3,7 +3,7 @@ import React from "react";
 import renderer from "react-test-renderer";
 import * as enzyme from "enzyme";
 import ReactSixteenAdapter from "enzyme-adapter-react-16";
-import FindPlacesCard from "../../../Components/Profile/Users/components/FindPlacesCard";
+import FindPlacesCard from "../../../Components/Users/FindPlacesCard";
 
 enzyme.configure({ adapter: new ReactSixteenAdapter() });
