+++ views/Settings/SettingsScreenStyles.js
